maven-plugin: Don't delete work directory in a certain case
... where it's equivalent or a sub-path of the META-INF/jacline
directory (which we can choose to delete or not).

+++ jacline-maven-plugin/src/main/java/com/kohlschutter/jacline/mavenplugin/JaclineCompileMojo.java
@@ -293,7 +293,12 @@ public class JaclineCompileMojo extends AbstractMojo {
       }
 
       if (Files.exists(jaclineWorkDirectoryPath)) {
-        IOUtil.deleteRecursively(jaclineWorkDirectoryPath);
+        if (jaclineWorkDirectoryPath.equals(jaclineMetaInfDirectoryPath) || jaclineWorkDirectoryPath
+            .startsWith(jaclineMetaInfDirectoryPath)) {
+          // handled above
+        } else {
+          IOUtil.deleteRecursively(jaclineWorkDirectoryPath);
+        }
       }
       Files.createDirectories(jaclineMetaInfDirectoryPath);
     } catch (IOException e) {
